refactor(#5175): split the authority contract and its tests out of two capped modules
The #5175 changes pushed both touched files past the 700-LoC namespace cap for
`src/services/discord/tmux_watcher/**` (`audit_maintainability` hard gate):
terminal_relay_plan.rs to 818 and turn_identity.rs to 767.

Move-only decomposition, no behavior change:
  * the soft-terminal authority contract (denial enum, the watcher turn-identity
    binding, and the pre-relay evaluation) moves to
    `tmux_watcher/turn_identity/soft_terminal_authority.rs`, re-exported from
    `turn_identity` so the watcher's existing glob import still resolves it;
  * the relay-plan authority tests move to `terminal_relay_plan_tests.rs` via
    `#[path]`, which also removes them from the production LoC the cap counts.

Test module paths are unchanged, so the curated lane and the lib inventory
manifest still name the same tests. Also pins the new curated invocation in the
`#4747` fast-check wiring contract, whose expected `test-non-pg` command list is
exact, and refreshes the generated ARCHITECTURE.md inventory for the two new
files.

Sizes after: terminal_relay_plan.rs 662, turn_identity.rs 597.
